fix(runtime): accept legacy repo backlog dispatch decisions
