+++ crates/harness-workflow/src/runtime/reducer.rs
@@ -1,21 +1,49 @@
+mod github_issue_completion;
+mod pr_feedback_completion;
+mod quality_gate_completion;
+mod repo_backlog_candidates;
+mod repo_backlog_completion;
+mod runtime_failure;
+mod support;
+
+use self::github_issue_completion::{
+    bind_pr_from_activity_result, closed_issue_evidence_from_activity_result,
+    closed_issue_evidence_from_activity_result_value, closed_issue_evidence_from_value,
+    github_issue_closed_decision, issue_implementation_missing_result_decision,
+};
+use self::pr_feedback_completion::{
+    pr_feedback_child_decision_from_activity_result,
+    pr_feedback_sweep_decision_from_activity_result,
+};
+use self::quality_gate_completion::{
+    quality_gate_activity_matches, quality_gate_success_contract_error,
+    quality_gate_success_decision,
+};
+use self::repo_backlog_completion::{
+    repo_backlog_child_dispatch_still_active, repo_backlog_invalid_success_decision,
+    repo_backlog_legacy_scan_dispatch_decision_from_activity_result,
+    repo_backlog_poll_decision_from_activity_result,
+    repo_backlog_sprint_plan_decision_from_activity_result,
+};
+use self::runtime_failure::{
+    retry_failed_activity_decision, runtime_blocked_decision, runtime_cancelled_decision,
+    runtime_failed_decision,
+};
+use self::support::{
+    event_command_type, invalid_agent_output_blocked_decision, runtime_completion_evidence,
+};
 use super::model::{
-    ActivityErrorKind, ActivityResult, ActivityStatus, WorkflowCommand, WorkflowCommandType,
-    WorkflowDecision, WorkflowEvent, WorkflowEvidence, WorkflowInstance,
+    ActivityResult, ActivityStatus, WorkflowCommand, WorkflowCommandType, WorkflowDecision,
+    WorkflowEvent, WorkflowInstance,
 };
-use super::pr_feedback::{
-    build_pr_feedback_decision, PrFeedbackDecisionInput, PrFeedbackOutcome,
-    PR_FEEDBACK_DEFINITION_ID, PR_FEEDBACK_INSPECT_ACTIVITY,
-};
+use super::pr_feedback::PR_FEEDBACK_DEFINITION_ID;
 use super::prompt_task::{PROMPT_TASK_DEFINITION_ID, PROMPT_TASK_IMPLEMENT_ACTIVITY};
-use super::quality_gate::{
-    QUALITY_BLOCKED_SIGNAL, QUALITY_FAILED_SIGNAL, QUALITY_GATE_ACTIVITY,
-    QUALITY_GATE_DEFINITION_ID, QUALITY_PASSED_SIGNAL,
-};
+use super::quality_gate::QUALITY_GATE_DEFINITION_ID;
 use super::repo_backlog::{
     REPO_BACKLOG_DEFINITION_ID, REPO_BACKLOG_POLL_ACTIVITY, REPO_BACKLOG_SPRINT_PLAN_ACTIVITY,
 };
 use super::validator::{DecisionValidator, ValidationContext};
-use chrono::{DateTime, Duration, Utc};
+use chrono::Utc;
 use serde_json::{json, Value};
 
 pub const RUNTIME_JOB_COMPLETED_EVENT: &str = "RuntimeJobCompleted";
@@ -67,7 +95,7 @@ fn reduce_success(
     event: &WorkflowEvent,
     result: &ActivityResult,
 ) -> Option<WorkflowDecision> {
-    let structured_decision = workflow_decision_from_activity_result(instance, event, result);
+    let structured_decision = workflow_decision_from_activity_result(event, result);
     if let Some(decision) = structured_decision
         .as_ref()
         .filter(|decision| structured_decision_validates(instance, event, result, decision))
@@ -133,6 +161,15 @@ fn reduce_success(
     if let Some(decision) =
         repo_backlog_sprint_plan_decision_from_activity_result(instance, event, result)
     {
+        return Some(decision);
+    }
+
+    if let Some(decision) = repo_backlog_legacy_scan_dispatch_decision_from_activity_result(
+        instance,
+        event,
+        result,
+        structured_decision.as_ref(),
+    ) {
         return Some(decision);
     }
 
@@ -202,7 +239,7 @@ fn reduce_success(
             "finish_repo_sprint_plan",
             "repo sprint planning completed without selected issue workflow commands",
         ),
-        (QUALITY_GATE_DEFINITION_ID, "checking", QUALITY_GATE_ACTIVITY) => (
+        (QUALITY_GATE_DEFINITION_ID, "checking", super::quality_gate::QUALITY_GATE_ACTIVITY) => (
             "passed",
             "quality_passed",
             "quality gate activity completed successfully",
@@ -236,105 +273,7 @@ fn reduce_success(
     Some(workflow_decision.high_confidence())
 }
 
-fn issue_implementation_missing_result_decision(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> Option<WorkflowDecision> {
-    if (
-        instance.definition_id.as_str(),
-        instance.state.as_str(),
-        result.activity.as_str(),
-    ) != (
-        GITHUB_ISSUE_PR_DEFINITION_ID,
-        "implementing",
-        "implement_issue",
-    ) {
-        return None;
-    }
-
-    let reason = "implement_issue succeeded without a pull_request artifact, closed-issue evidence, or another validated terminal signal";
-    Some(
-        WorkflowDecision::new(
-            &instance.id,
-            &instance.state,
-            "block_missing_implementation_result",
-            "blocked",
-            reason,
-        )
-        .with_command(WorkflowCommand::mark_blocked(
-            reason,
-            format!(
-                "runtime-completion:{}:missing-implementation:block",
-                event.id
-            ),
-        ))
-        .with_command(WorkflowCommand::new(
-            WorkflowCommandType::RequestOperatorAttention,
-            format!(
-                "runtime-completion:{}:missing-implementation:operator",
-                event.id
-            ),
-            json!({
-                "reason": reason,
-                "activity": result.activity,
-                "runtime_job_id": event_field_string(event, "runtime_job_id"),
-            }),
-        ))
-        .with_evidence(runtime_completion_evidence(event, result))
-        .high_confidence(),
-    )
-}
-
-fn github_issue_closed_decision(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> Option<WorkflowDecision> {
-    if instance.definition_id != GITHUB_ISSUE_PR_DEFINITION_ID
-        || !github_issue_state_can_finish_closed(instance.state.as_str())
-    {
-        return None;
-    }
-
-    let closed_issue = closed_issue_evidence_from_activity_result(result)?;
-    let reason = format!(
-        "{} reported structured evidence that the GitHub issue is already closed",
-        result.activity
-    );
-    Some(
-        WorkflowDecision::new(
-            &instance.id,
-            &instance.state,
-            "finish_closed_issue",
-            "done",
-            &reason,
-        )
-        .with_command(WorkflowCommand::new(
-            WorkflowCommandType::MarkDone,
-            format!("runtime-completion:{}:closed-issue:done", event.id),
-            json!({
-                "reason": reason,
-                "activity": result.activity,
-                "runtime_job_id": event_field_string(event, "runtime_job_id"),
-                "closed_issue_evidence": closed_issue.payload,
-            }),
-        ))
-        .with_evidence(WorkflowEvidence::new("closed_issue", closed_issue.summary))
-        .with_evidence(runtime_completion_evidence(event, result))
-        .high_confidence(),
-    )
-}
-
-fn github_issue_state_can_finish_closed(state: &str) -> bool {
-    matches!(
-        state,
-        "implementing" | "pr_open" | "awaiting_feedback" | "addressing_feedback" | "ready_to_merge"
-    )
-}
-
 fn workflow_decision_from_activity_result(
-    _instance: &WorkflowInstance,
     event: &WorkflowEvent,
     result: &ActivityResult,
 ) -> Option<WorkflowDecision> {
@@ -346,675 +285,6 @@ fn workflow_decision_from_activity_result(
             serde_json::from_value::<WorkflowDecision>(artifact.artifact.clone()).ok()
         })
         .map(|decision| decision.with_evidence(runtime_completion_evidence(event, result)))
-}
-
-fn repo_backlog_poll_decision_from_activity_result(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> Option<WorkflowDecision> {
-    if (
-        instance.definition_id.as_str(),
-        instance.state.as_str(),
-        result.activity.as_str(),
-    ) != (
-        REPO_BACKLOG_DEFINITION_ID,
-        "scanning",
-        REPO_BACKLOG_POLL_ACTIVITY,
-    ) {
-        return None;
-    }
-
-    let parent_repo = optional_data_string(instance, "repo");
-    let discovered =
-        issue_candidates_from_signals(result, "IssueDiscovered", parent_repo.as_deref());
-    let pr_feedback = pr_feedback_candidates_from_signals(
-        result,
-        "OpenPrFeedbackDiscovered",
-        parent_repo.as_deref(),
-    );
-    let known_dependencies =
-        known_dependency_candidates_from_skipped_signals(result, parent_repo.as_deref());
-
-    if discovered.is_empty() && pr_feedback.is_empty() {
-        return None;
-    }
-
-    if discovered.is_empty() {
-        let mut decision = WorkflowDecision::new(
-            &instance.id,
-            &instance.state,
-            "start_open_pr_feedback_tasks_from_scan",
-            "dispatching",
-            format!(
-                "repo backlog scan discovered {} open PR(s) with actionable review feedback",
-                pr_feedback.len()
-            ),
-        )
-        .with_evidence(runtime_completion_evidence(event, result));
-
-        for candidate in pr_feedback {
-            decision = decision
-                .with_command(candidate.start_prompt_task_command(event))
-                .with_evidence(candidate.github_pr_evidence());
-        }
-
-        return Some(decision.high_confidence());
-    }
-
-    let issue_payloads = discovered
-        .iter()
-        .map(RepoBacklogIssueCandidate::to_command_value)
-        .collect::<Vec<_>>();
-    let pr_feedback_payloads = pr_feedback
-        .iter()
-        .map(RepoBacklogPrFeedbackCandidate::to_command_value)
-        .collect::<Vec<_>>();
-    let known_dependency_payloads = known_dependencies
-        .iter()
-        .map(RepoBacklogIssueCandidate::to_command_value)
-        .collect::<Vec<_>>();
-    let mut decision = WorkflowDecision::new(
-        &instance.id,
-        &instance.state,
-        "plan_repo_sprint_from_scan",
-        "planning_batch",
-        format!(
-            "repo backlog scan discovered {} open issue(s); runtime sprint planning must select dispatch order",
-            discovered.len()
-        ),
-    )
-    .with_command(WorkflowCommand::new(
-        WorkflowCommandType::EnqueueActivity,
-        format!("runtime-completion:{}:plan-repo-sprint", event.id),
-        json!({
-            "activity": REPO_BACKLOG_SPRINT_PLAN_ACTIVITY,
-            "repo": parent_repo,
-            "issues": issue_payloads,
-            "open_pr_feedback": pr_feedback_payloads,
-            "known_dependencies": known_dependency_payloads,
-        }),
-    ))
-    .with_evidence(runtime_completion_evidence(event, result));
-
-    for candidate in discovered {
-        decision = decision.with_evidence(candidate.github_issue_evidence());
-    }
-    for candidate in pr_feedback {
-        decision = decision.with_evidence(candidate.github_pr_evidence());
-    }
-
-    Some(decision.high_confidence())
-}
-
-fn repo_backlog_sprint_plan_decision_from_activity_result(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> Option<WorkflowDecision> {
-    if (
-        instance.definition_id.as_str(),
-        instance.state.as_str(),
-        result.activity.as_str(),
-    ) != (
-        REPO_BACKLOG_DEFINITION_ID,
-        "planning_batch",
-        REPO_BACKLOG_SPRINT_PLAN_ACTIVITY,
-    ) {
-        return None;
-    }
-
-    let parent_repo = optional_data_string(instance, "repo");
-    let selected = normalize_selected_sprint_dependencies(
-        sprint_plan_selected_issues(result, event, parent_repo.as_deref()),
-        &sprint_plan_known_dependency_inputs(event, parent_repo.as_deref()),
-    );
-    let pr_feedback = sprint_plan_pr_feedback_inputs(event, parent_repo.as_deref());
-    if selected.is_empty() && pr_feedback.is_empty() {
-        return None;
-    }
-
-    let decision_name = if pr_feedback.is_empty() {
-        "start_issue_workflows_from_sprint_plan"
-    } else {
-        "start_repo_backlog_workflows_from_sprint_plan"
-    };
-    let reason = if pr_feedback.is_empty() {
-        format!(
-            "runtime sprint planner selected {} issue workflow(s) for dispatch",
-            selected.len()
-        )
-    } else {
-        format!(
-            "runtime sprint planner selected {} issue workflow(s) and {} open PR feedback task(s) for dispatch",
-            selected.len(),
-            pr_feedback.len()
-        )
-    };
-    let mut decision = WorkflowDecision::new(
-        &instance.id,
-        &instance.state,
-        decision_name,
-        "dispatching",
-        reason,
-    )
-    .with_evidence(runtime_completion_evidence(event, result));
-
-    for candidate in selected {
-        let issue_number = candidate.issue_number;
-        let repo_key = candidate.repo.as_deref().unwrap_or("<none>");
-        decision = decision
-            .with_command(WorkflowCommand::new(
-                WorkflowCommandType::StartChildWorkflow,
-                format!("repo-sprint-plan:{repo_key}:issue:{issue_number}:start"),
-                json!({
-                    "definition_id": "github_issue_pr",
-                    "subject_key": format!("issue:{issue_number}"),
-                    "repo": candidate.repo,
-                    "issue_number": issue_number,
-                    "issue_url": candidate.issue_url,
-                    "title": candidate.title,
-                    "labels": candidate.labels,
-                    "depends_on": candidate.depends_on,
-                    "source": "github",
-                    "external_id": issue_number.to_string(),
-                    "auto_submit": true,
-                }),
-            ))
-            .with_evidence(candidate.github_issue_evidence());
-    }
-    for candidate in pr_feedback {
-        decision = decision
-            .with_command(candidate.start_prompt_task_command(event))
-            .with_evidence(candidate.github_pr_evidence());
-    }
-
-    Some(decision.high_confidence())
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct RepoBacklogIssueCandidate {
-    issue_number: u64,
-    repo: Option<String>,
-    issue_url: Option<String>,
-    title: Option<String>,
-    labels: Vec<String>,
-    depends_on: Vec<u64>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct RepoBacklogPrFeedbackCandidate {
-    pr_number: u64,
-    repo: Option<String>,
-    pr_url: Option<String>,
-    title: Option<String>,
-    feedback_count: Option<u64>,
-    summary: Option<String>,
-}
-
-impl RepoBacklogPrFeedbackCandidate {
-    fn from_value(value: &Value, parent_repo: Option<&str>) -> Option<Self> {
-        let pr_number = value
-            .get("pr_number")
-            .or_else(|| value.get("pull_request"))
-            .or_else(|| value.get("pr"))
-            .and_then(json_value_u64)?;
-        let repo = value
-            .get("repo")
-            .and_then(non_empty_json_string)
-            .or_else(|| parent_repo.map(ToOwned::to_owned));
-        let pr_url = value
-            .get("pr_url")
-            .or_else(|| value.get("pull_request_url"))
-            .and_then(non_empty_json_string);
-        let title = value.get("title").and_then(non_empty_json_string);
-        let feedback_count = value
-            .get("feedback_count")
-            .or_else(|| value.get("review_thread_count"))
-            .or_else(|| value.get("thread_count"))
-            .and_then(json_value_u64);
-        let summary = value
-            .get("summary")
-            .or_else(|| value.get("reason"))
-            .and_then(non_empty_json_string);
-        Some(Self {
-            pr_number,
-            repo,
-            pr_url,
-            title,
-            feedback_count,
-            summary,
-        })
-    }
-
-    fn start_prompt_task_command(&self, event: &WorkflowEvent) -> WorkflowCommand {
-        let repo_key = self.repo.as_deref().unwrap_or("<none>");
-        let external_id = format!("pr-feedback:{repo_key}:{}", self.pr_number);
-        WorkflowCommand::new(
-            WorkflowCommandType::StartChildWorkflow,
-            format!("repo-backlog:{repo_key}:pr:{}:feedback", self.pr_number),
-            json!({
-                "definition_id": PROMPT_TASK_DEFINITION_ID,
-                "subject_key": format!("pr:{}:feedback", self.pr_number),
-                "repo": self.repo,
-                "pr_number": self.pr_number,
-                "pr_url": self.pr_url,
-                "title": self.title,
-                "feedback_count": self.feedback_count,
-                "source": "github_pr_feedback",
-                "external_id": external_id,
-                "task_id": format!("repo-backlog:{repo_key}:pr:{}:feedback", self.pr_number),
-                "prompt": self.prompt(),
-                "source_event_id": event.id,
-            }),
-        )
-    }
-
-    fn to_command_value(&self) -> Value {
-        json!({
-            "pr_number": self.pr_number,
-            "repo": self.repo,
-            "pr_url": self.pr_url,
-            "title": self.title,
-            "feedback_count": self.feedback_count,
-            "summary": self.summary,
-        })
-    }
-
-    fn prompt(&self) -> String {
-        let repo = self.repo.as_deref().unwrap_or("<unknown repo>");
-        let pr_ref = self
-            .pr_url
-            .as_deref()
-            .map(ToOwned::to_owned)
-            .unwrap_or_else(|| format!("{repo}#{}", self.pr_number));
-        let title = self.title.as_deref().unwrap_or("<unknown title>");
-        let feedback_count = self
-            .feedback_count
-            .map(|count| count.to_string())
-            .unwrap_or_else(|| "<unknown>".to_string());
-        let summary = self.summary.as_deref().unwrap_or("No summary provided.");
-        format!(
-            "Handle unresolved review feedback for PR {pr_ref}.\n\n\
-Repository: {repo}\n\
-PR number: {}\n\
-PR title: {title}\n\
-Unresolved/actionable feedback count: {feedback_count}\n\
-Discovery summary: {summary}\n\n\
-Inspect the PR review threads, CI state, and mergeability. Address valid requested changes on the PR branch, run the relevant validation commands, push the branch, and reply to or resolve addressed review threads when possible. Do not make unrelated changes.",
-            self.pr_number
-        )
-    }
-
-    fn github_pr_evidence(&self) -> WorkflowEvidence {
-        let repo = self.repo.as_deref().unwrap_or("<none>");
-        let url = self.pr_url.as_deref().unwrap_or("<unknown>");
-        WorkflowEvidence::new(
-            "github_pr_feedback",
-            format!("repo={repo} pr={} url={url}", self.pr_number),
-        )
-    }
-}
-
-impl RepoBacklogIssueCandidate {
-    fn from_value(value: &Value, parent_repo: Option<&str>) -> Option<Self> {
-        let issue_number = value
-            .get("issue_number")
-            .or_else(|| value.get("issue"))
-            .and_then(json_value_u64)?;
-        let repo = value
-            .get("repo")
-            .and_then(non_empty_json_string)
-            .or_else(|| parent_repo.map(ToOwned::to_owned));
-        let issue_url = value.get("issue_url").and_then(non_empty_json_string);
-        let title = value.get("title").and_then(non_empty_json_string);
-        let labels = string_array_field(value, "labels");
-        let depends_on = u64_array_field(value, "depends_on");
-        Some(Self {
-            issue_number,
-            repo,
-            issue_url,
-            title,
-            labels,
-            depends_on,
-        })
-    }
-
-    fn from_sprint_task(
-        value: &Value,
-        candidates: &[Self],
-        parent_repo: Option<&str>,
-    ) -> Option<Self> {
-        let issue_number = value
-            .get("issue_number")
-            .or_else(|| value.get("issue"))
-            .and_then(json_value_u64)?;
-        let mut candidate = candidates
-            .iter()
-            .find(|candidate| candidate.issue_number == issue_number)
-            .cloned()
-            .unwrap_or(Self {
-                issue_number,
-                repo: parent_repo.map(ToOwned::to_owned),
-                issue_url: None,
-                title: None,
-                labels: Vec::new(),
-                depends_on: Vec::new(),
-            });
-        if let Some(repo) = value.get("repo").and_then(non_empty_json_string) {
-            candidate.repo = Some(repo);
-        }
-        if let Some(issue_url) = value.get("issue_url").and_then(non_empty_json_string) {
-            candidate.issue_url = Some(issue_url);
-        }
-        if let Some(title) = value.get("title").and_then(non_empty_json_string) {
-            candidate.title = Some(title);
-        }
-        let labels = string_array_field(value, "labels");
-        if !labels.is_empty() {
-            candidate.labels = labels;
-        }
-        if value.get("depends_on").is_some() {
-            candidate.depends_on = u64_array_field(value, "depends_on");
-        }
-        Some(candidate)
-    }
-
-    fn to_command_value(&self) -> Value {
-        json!({
-            "issue_number": self.issue_number,
-            "repo": self.repo,
-            "issue_url": self.issue_url,
-            "title": self.title,
-            "labels": self.labels,
-            "depends_on": self.depends_on,
-        })
-    }
-
-    fn merge_from(&mut self, other: Self) {
-        if self.repo.is_none() {
-            self.repo = other.repo;
-        }
-        if self.issue_url.is_none() {
-            self.issue_url = other.issue_url;
-        }
-        if self.title.is_none() {
-            self.title = other.title;
-        }
-        append_missing_strings(&mut self.labels, other.labels);
-        append_missing_u64s(&mut self.depends_on, other.depends_on);
-    }
-
-    fn github_issue_evidence(&self) -> WorkflowEvidence {
-        let repo_key = self.repo.as_deref().unwrap_or("<none>");
-        WorkflowEvidence::new(
-            "github_issue",
-            match self.issue_url.as_deref() {
-                Some(url) => format!("repo={repo_key} issue={} url={url}", self.issue_number),
-                None => format!("repo={repo_key} issue={}", self.issue_number),
-            },
-        )
-    }
-}
-
-fn issue_candidates_from_signals(
-    result: &ActivityResult,
-    signal_type: &str,
-    parent_repo: Option<&str>,
-) -> Vec<RepoBacklogIssueCandidate> {
-    result
-        .signals
-        .iter()
-        .filter(|signal| signal.signal_type == signal_type)
-        .filter_map(|signal| RepoBacklogIssueCandidate::from_value(&signal.signal, parent_repo))
-        .collect()
-}
-
-fn pr_feedback_candidates_from_signals(
-    result: &ActivityResult,
-    signal_type: &str,
-    parent_repo: Option<&str>,
-) -> Vec<RepoBacklogPrFeedbackCandidate> {
-    result
-        .signals
-        .iter()
-        .filter(|signal| signal.signal_type == signal_type)
-        .filter_map(|signal| {
-            RepoBacklogPrFeedbackCandidate::from_value(&signal.signal, parent_repo)
-        })
-        .collect()
-}
-
-fn known_dependency_candidates_from_skipped_signals(
-    result: &ActivityResult,
-    parent_repo: Option<&str>,
-) -> Vec<RepoBacklogIssueCandidate> {
-    result
-        .signals
-        .iter()
-        .filter(|signal| signal.signal_type == "IssueSkipped")
-        .filter(|signal| skipped_issue_has_known_workflow(&signal.signal))
-        .filter_map(|signal| RepoBacklogIssueCandidate::from_value(&signal.signal, parent_repo))
-        .collect()
-}
-
-fn skipped_issue_has_known_workflow(value: &Value) -> bool {
-    value
-        .get("existing_workflow")
-        .and_then(Value::as_bool)
-        .unwrap_or(false)
-        || value
-            .get("has_workflow")
-            .and_then(Value::as_bool)
-            .unwrap_or(false)
-        || value
-            .get("workflow_state")
-            .and_then(non_empty_json_string)
-            .is_some()
-        || value
-            .get("runtime_workflow_state")
-            .and_then(non_empty_json_string)
-            .is_some()
-        || value
-            .get("issue_workflow_state")
-            .and_then(non_empty_json_string)
-            .is_some()
-}
-
-fn sprint_plan_selected_issues(
-    result: &ActivityResult,
-    event: &WorkflowEvent,
-    parent_repo: Option<&str>,
-) -> Vec<RepoBacklogIssueCandidate> {
-    let mut scan_candidates = sprint_plan_candidate_inputs(event, parent_repo);
-    scan_candidates.extend(sprint_plan_candidate_artifacts(result, parent_repo));
-    let mut selected = sprint_plan_selected_signals(result, &scan_candidates, parent_repo);
-    selected.extend(sprint_plan_selected_artifacts(
-        result,
-        &scan_candidates,
-        parent_repo,
-    ));
-    dedupe_issue_candidates(selected)
-}
-
-fn sprint_plan_known_dependency_inputs(
-    event: &WorkflowEvent,
-    parent_repo: Option<&str>,
-) -> Vec<RepoBacklogIssueCandidate> {
-    event_workflow_command(event)
-        .and_then(|command| {
-            command
-                .command
-                .get("known_dependencies")
-                .and_then(Value::as_array)
-                .map(|items| {
-                    items
-                        .iter()
-                        .filter_map(|value| {
-                            RepoBacklogIssueCandidate::from_value(value, parent_repo)
-                        })
-                        .collect()
-                })
-        })
-        .unwrap_or_default()
-}
-
-fn sprint_plan_pr_feedback_inputs(
-    event: &WorkflowEvent,
-    parent_repo: Option<&str>,
-) -> Vec<RepoBacklogPrFeedbackCandidate> {
-    event_workflow_command(event)
-        .and_then(|command| {
-            command
-                .command
-                .get("open_pr_feedback")
-                .and_then(Value::as_array)
-                .map(|items| {
-                    items
-                        .iter()
-                        .filter_map(|value| {
-                            RepoBacklogPrFeedbackCandidate::from_value(value, parent_repo)
-                        })
-                        .collect()
-                })
-        })
-        .unwrap_or_default()
-}
-
-fn sprint_plan_candidate_inputs(
-    event: &WorkflowEvent,
-    parent_repo: Option<&str>,
-) -> Vec<RepoBacklogIssueCandidate> {
-    event_workflow_command(event)
-        .and_then(|command| {
-            command
-                .command
-                .get("issues")
-                .and_then(Value::as_array)
-                .map(|items| {
-                    items
-                        .iter()
-                        .filter_map(|value| {
-                            RepoBacklogIssueCandidate::from_value(value, parent_repo)
-                        })
-                        .collect()
-                })
-        })
-        .unwrap_or_default()
-}
-
-fn sprint_plan_candidate_artifacts(
-    result: &ActivityResult,
-    parent_repo: Option<&str>,
-) -> Vec<RepoBacklogIssueCandidate> {
-    result
-        .artifacts
-        .iter()
-        .filter(|artifact| artifact.artifact_type == "repo_backlog_candidates")
-        .flat_map(|artifact| array_items(&artifact.artifact, "issues"))
-        .filter_map(|value| RepoBacklogIssueCandidate::from_value(value, parent_repo))
-        .collect()
-}
-
-fn sprint_plan_selected_signals(
-    result: &ActivityResult,
-    candidates: &[RepoBacklogIssueCandidate],
-    parent_repo: Option<&str>,
-) -> Vec<RepoBacklogIssueCandidate> {
-    result
-        .signals
-        .iter()
-        .filter(|signal| signal.signal_type == "SprintTaskSelected")
-        .filter_map(|signal| {
-            RepoBacklogIssueCandidate::from_sprint_task(&signal.signal, candidates, parent_repo)
-        })
-        .collect()
-}
-
-fn sprint_plan_selected_artifacts(
-    result: &ActivityResult,
-    candidates: &[RepoBacklogIssueCandidate],
-    parent_repo: Option<&str>,
-) -> Vec<RepoBacklogIssueCandidate> {
-    result
-        .artifacts
-        .iter()
-        .filter(|artifact| artifact.artifact_type == "sprint_plan")
-        .flat_map(|artifact| array_items(&artifact.artifact, "tasks"))
-        .filter_map(|value| {
-            RepoBacklogIssueCandidate::from_sprint_task(value, candidates, parent_repo)
-        })
-        .collect()
-}
-
-fn dedupe_issue_candidates(
-    candidates: Vec<RepoBacklogIssueCandidate>,
-) -> Vec<RepoBacklogIssueCandidate> {
-    let mut index_by_key = std::collections::BTreeMap::<(String, u64), usize>::new();
-    let mut merged = Vec::<RepoBacklogIssueCandidate>::new();
-    for candidate in candidates {
-        let key = (
-            candidate.repo.clone().unwrap_or_default(),
-            candidate.issue_number,
-        );
-        if let Some(index) = index_by_key.get(&key).copied() {
-            merged[index].merge_from(candidate);
-        } else {
-            index_by_key.insert(key, merged.len());
-            merged.push(candidate);
-        }
-    }
-    merged
-}
-
-fn normalize_selected_sprint_dependencies(
-    mut candidates: Vec<RepoBacklogIssueCandidate>,
-    known_dependencies: &[RepoBacklogIssueCandidate],
-) -> Vec<RepoBacklogIssueCandidate> {
-    let mut valid_dependencies: std::collections::BTreeSet<_> = candidates
-        .iter()
-        .map(|candidate| {
-            (
-                candidate.repo.clone().unwrap_or_default(),
-                candidate.issue_number,
-            )
-        })
-        .collect();
-    valid_dependencies.extend(known_dependencies.iter().map(|candidate| {
-        (
-            candidate.repo.clone().unwrap_or_default(),
-            candidate.issue_number,
-        )
-    }));
-
-    for candidate in &mut candidates {
-        let repo = candidate.repo.clone().unwrap_or_default();
-        let mut seen = std::collections::BTreeSet::new();
-        let issue_number = candidate.issue_number;
-        candidate.depends_on.retain(|dependency| {
-            *dependency != issue_number
-                && valid_dependencies.contains(&(repo.clone(), *dependency))
-                && seen.insert(*dependency)
-        });
-    }
-
-    candidates
-}
-
-fn append_missing_strings(target: &mut Vec<String>, values: Vec<String>) {
-    for value in values {
-        if !target.contains(&value) {
-            target.push(value);
-        }
-    }
-}
-
-fn append_missing_u64s(target: &mut Vec<u64>, values: Vec<u64>) {
-    for value in values {
-        if !target.contains(&value) {
-            target.push(value);
-        }
-    }
 }
 
 fn structured_decision_validates(
@@ -1052,1008 +322,4 @@ fn structured_decision_validates(
             &ValidationContext::new(event.source.as_str(), Utc::now()),
         )
         .is_ok()
-}
-
-fn repo_backlog_invalid_success_decision(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-    structured_decision: Option<&WorkflowDecision>,
-) -> Option<WorkflowDecision> {
-    match (
-        instance.definition_id.as_str(),
-        instance.state.as_str(),
-        result.activity.as_str(),
-    ) {
-        (REPO_BACKLOG_DEFINITION_ID, "scanning", REPO_BACKLOG_POLL_ACTIVITY) => {
-            let issue_discovered_count = signal_count(result, "IssueDiscovered");
-            let valid_issue_discovered_count =
-                issue_candidates_from_signals(result, "IssueDiscovered", None).len();
-            if issue_discovered_count > 0 && valid_issue_discovered_count == 0 {
-                return Some(invalid_agent_output_blocked_decision(
-                    instance,
-                    event,
-                    result,
-                    "repo backlog poll emitted IssueDiscovered signals, but none contained a valid issue_number",
-                ));
-            }
-            let pr_feedback_count = signal_count(result, "OpenPrFeedbackDiscovered");
-            let valid_pr_feedback_count =
-                pr_feedback_candidates_from_signals(result, "OpenPrFeedbackDiscovered", None).len();
-            if pr_feedback_count > 0 && valid_pr_feedback_count == 0 {
-                return Some(invalid_agent_output_blocked_decision(
-                    instance,
-                    event,
-                    result,
-                    "repo backlog poll emitted OpenPrFeedbackDiscovered signals, but none contained a valid pr_number",
-                ));
-            }
-            if !has_any_signal(
-                result,
-                &[
-                    "IssueDiscovered",
-                    "IssueSkipped",
-                    "NoOpenIssueFound",
-                    "OpenPrFeedbackDiscovered",
-                    "OpenPrFeedbackSkipped",
-                    "NoOpenPrFeedbackFound",
-                ],
-            ) {
-                return Some(invalid_agent_output_blocked_decision(
-                    instance,
-                    event,
-                    result,
-                    "repo backlog poll succeeded without issue or open PR feedback signals",
-                ));
-            }
-            if structured_decision.is_some() {
-                return Some(invalid_agent_output_blocked_decision(
-                    instance,
-                    event,
-                    result,
-                    "repo backlog poll emitted a workflow_decision artifact that did not validate and no valid domain fallback was available",
-                ));
-            }
-            None
-        }
-        (REPO_BACKLOG_DEFINITION_ID, "planning_batch", REPO_BACKLOG_SPRINT_PLAN_ACTIVITY) => {
-            let sprint_task_selected_count = signal_count(result, "SprintTaskSelected");
-            let sprint_plan_task_count = sprint_plan_task_count(result);
-            let has_valid_noop_sprint_plan_artifact = has_valid_noop_sprint_plan_artifact(result);
-            if (sprint_task_selected_count > 0 || sprint_plan_task_count > 0)
-                && sprint_plan_selected_issues(result, event, None).is_empty()
-            {
-                return Some(invalid_agent_output_blocked_decision(
-                    instance,
-                    event,
-                    result,
-                    "repo sprint plan emitted selected task output, but no selected issue had a valid issue_number",
-                ));
-            }
-            if !has_any_signal(
-                result,
-                &["SprintTaskSelected", "IssueSkipped", "NoSprintTaskSelected"],
-            ) && sprint_plan_task_count == 0
-                && !has_valid_noop_sprint_plan_artifact
-            {
-                return Some(invalid_agent_output_blocked_decision(
-                    instance,
-                    event,
-                    result,
-                    "repo sprint plan succeeded without SprintTaskSelected, IssueSkipped, NoSprintTaskSelected, or a valid no-op sprint_plan artifact",
-                ));
-            }
-            if structured_decision.is_some() {
-                return Some(invalid_agent_output_blocked_decision(
-                    instance,
-                    event,
-                    result,
-                    "repo sprint plan emitted a workflow_decision artifact that did not validate and no valid domain fallback was available",
-                ));
-            }
-            None
-        }
-        _ => None,
-    }
-}
-
-fn invalid_agent_output_blocked_decision(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-    reason: &str,
-) -> WorkflowDecision {
-    WorkflowDecision::new(
-        &instance.id,
-        &instance.state,
-        "block_invalid_agent_output",
-        "blocked",
-        reason,
-    )
-    .with_command(WorkflowCommand::mark_blocked(
-        reason,
-        format!("runtime-completion:{}:invalid-output:block", event.id),
-    ))
-    .with_command(WorkflowCommand::new(
-        WorkflowCommandType::RequestOperatorAttention,
-        format!("runtime-completion:{}:invalid-output:operator", event.id),
-        json!({
-            "reason": reason,
-            "activity": result.activity,
-            "runtime_job_id": event_field_string(event, "runtime_job_id"),
-        }),
-    ))
-    .with_evidence(runtime_completion_evidence(event, result))
-    .high_confidence()
-}
-
-fn quality_gate_success_decision(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> Option<WorkflowDecision> {
-    if !quality_gate_activity_matches(instance, result) {
-        return None;
-    }
-
-    if quality_gate_success_contract_error(result).is_none() {
-        return Some(
-            WorkflowDecision::new(
-                &instance.id,
-                &instance.state,
-                "quality_passed",
-                "passed",
-                "quality gate activity completed successfully with passing evidence",
-            )
-            .with_evidence(runtime_completion_evidence(event, result))
-            .high_confidence(),
-        );
-    }
-
-    let reason = quality_gate_success_contract_error(result)?;
-    Some(invalid_agent_output_blocked_decision(
-        instance, event, result, reason,
-    ))
-}
-
-fn quality_gate_activity_matches(instance: &WorkflowInstance, result: &ActivityResult) -> bool {
-    (
-        instance.definition_id.as_str(),
-        instance.state.as_str(),
-        result.activity.as_str(),
-    ) == (
-        QUALITY_GATE_DEFINITION_ID,
-        "checking",
-        QUALITY_GATE_ACTIVITY,
-    )
-}
-
-fn quality_gate_success_contract_error(result: &ActivityResult) -> Option<&'static str> {
-    let passed = signal_count(result, QUALITY_PASSED_SIGNAL);
-    let failed = signal_count(result, QUALITY_FAILED_SIGNAL);
-    let blocked = signal_count(result, QUALITY_BLOCKED_SIGNAL);
-    let status_count = passed + failed + blocked;
-    if status_count == 0 {
-        Some("run_quality_gate succeeded without a quality status signal")
-    } else if passed == 0 {
-        Some("run_quality_gate succeeded without a QualityPassed signal")
-    } else if passed > 1 || failed > 0 || blocked > 0 {
-        Some("run_quality_gate succeeded with ambiguous quality status signals")
-    } else if !quality_gate_has_validation_evidence(result) {
-        Some("run_quality_gate succeeded without validation evidence")
-    } else {
-        None
-    }
-}
-
-fn quality_gate_has_validation_evidence(result: &ActivityResult) -> bool {
-    result
-        .validation
-        .iter()
-        .any(|record| !record.command.trim().is_empty())
-        || result
-            .artifacts
-            .iter()
-            .any(|artifact| artifact.artifact_type == "validation_report")
-}
-
-fn signal_count(result: &ActivityResult, signal_type: &str) -> usize {
-    result
-        .signals
-        .iter()
-        .filter(|signal| signal.signal_type == signal_type)
-        .count()
-}
-
-fn has_any_signal(result: &ActivityResult, signal_types: &[&str]) -> bool {
-    result
-        .signals
-        .iter()
-        .any(|signal| signal_types.contains(&signal.signal_type.as_str()))
-}
-
-fn sprint_plan_task_count(result: &ActivityResult) -> usize {
-    result
-        .artifacts
-        .iter()
-        .filter(|artifact| artifact.artifact_type == "sprint_plan")
-        .map(|artifact| array_items(&artifact.artifact, "tasks").len())
-        .sum()
-}
-
-fn has_valid_noop_sprint_plan_artifact(result: &ActivityResult) -> bool {
-    result
-        .artifacts
-        .iter()
-        .filter(|artifact| artifact.artifact_type == "sprint_plan")
-        .any(|artifact| valid_noop_sprint_plan_artifact(&artifact.artifact))
-}
-
-fn valid_noop_sprint_plan_artifact(value: &Value) -> bool {
-    let Some(tasks) = value.get("tasks").and_then(Value::as_array) else {
-        return false;
-    };
-    tasks.is_empty() && (has_valid_skip_evidence(value) || has_explicit_noop_evidence(value))
-}
-
-fn has_valid_skip_evidence(value: &Value) -> bool {
-    value
-        .get("skip")
-        .or_else(|| value.get("skipped"))
-        .and_then(Value::as_array)
-        .is_some_and(|items| items.iter().any(valid_sprint_plan_skip_evidence))
-}
-
-fn valid_sprint_plan_skip_evidence(value: &Value) -> bool {
-    let has_issue = value
-        .get("issue")
-        .or_else(|| value.get("issue_number"))
-        .and_then(json_value_u64)
-        .is_some_and(|issue_number| issue_number > 0);
-    let has_reason = value
-        .get("reason")
-        .or_else(|| value.get("note"))
-        .and_then(non_empty_json_string)
-        .is_some();
-    has_issue && has_reason
-}
-
-fn has_explicit_noop_evidence(value: &Value) -> bool {
-    let has_noop_flag = value
-        .get("no_task_selected")
-        .or_else(|| value.get("no_tasks_selected"))
-        .or_else(|| value.get("no_sprint_task_selected"))
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    let has_reason = value
-        .get("reason")
-        .or_else(|| value.get("summary"))
-        .or_else(|| value.get("note"))
-        .and_then(non_empty_json_string)
-        .is_some();
-    has_noop_flag && has_reason
-}
-
-fn repo_backlog_child_dispatch_still_active(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-) -> bool {
-    instance.definition_id == REPO_BACKLOG_DEFINITION_ID
-        && instance.state == "dispatching"
-        && event_command_type(event) == Some("start_child_workflow")
-        && event
-            .event
-            .get("active_start_child_workflow_commands")
-            .and_then(Value::as_u64)
-            .unwrap_or(1)
-            > 0
-}
-
-fn bind_pr_from_activity_result(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> Option<WorkflowDecision> {
-    if (
-        instance.definition_id.as_str(),
-        instance.state.as_str(),
-        result.activity.as_str(),
-    ) != (
-        GITHUB_ISSUE_PR_DEFINITION_ID,
-        "implementing",
-        "implement_issue",
-    ) {
-        return None;
-    }
-    let (pr_number, pr_url) = pull_request_artifact(result)?;
-    Some(
-        WorkflowDecision::new(
-            &instance.id,
-            &instance.state,
-            "bind_pr",
-            "pr_open",
-            "implementation activity returned a structured pull request artifact",
-        )
-        .with_command(WorkflowCommand::bind_pr(
-            pr_number,
-            pr_url.clone(),
-            format!("runtime-completion:{}:bind-pr:{pr_number}", event.id),
-        ))
-        .with_evidence(WorkflowEvidence::new("pull_request", pr_url))
-        .with_evidence(runtime_completion_evidence(event, result))
-        .high_confidence(),
-    )
-}
-
-fn pull_request_artifact(result: &ActivityResult) -> Option<(u64, String)> {
-    result
-        .artifacts
-        .iter()
-        .filter(|artifact| artifact.artifact_type == "pull_request")
-        .find_map(|artifact| {
-            let pr_number = artifact.artifact.get("pr_number")?.as_u64()?;
-            let pr_url = artifact
-                .artifact
-                .get("pr_url")?
-                .as_str()
-                .filter(|value| !value.trim().is_empty())?
-                .to_string();
-            Some((pr_number, pr_url))
-        })
-}
-
-#[derive(Debug, Clone)]
-struct ClosedIssueEvidence {
-    summary: String,
-    payload: Value,
-}
-
-fn closed_issue_evidence_from_activity_result(
-    result: &ActivityResult,
-) -> Option<ClosedIssueEvidence> {
-    result
-        .signals
-        .iter()
-        .find_map(|signal| match signal.signal_type.as_str() {
-            ISSUE_CLOSED_SIGNAL | ISSUE_ALREADY_RESOLVED_SIGNAL => {
-                closed_issue_evidence_from_value(&signal.signal, &signal.signal_type)
-            }
-            _ => None,
-        })
-        .or_else(|| {
-            result
-                .artifacts
-                .iter()
-                .filter(|artifact| artifact.artifact_type == ISSUE_STATE_ARTIFACT)
-                .find_map(|artifact| {
-                    if issue_state_is_closed(&artifact.artifact) {
-                        closed_issue_evidence_from_value(&artifact.artifact, ISSUE_STATE_ARTIFACT)
-                    } else {
-                        None
-                    }
-                })
-        })
-}
-
-fn closed_issue_evidence_from_activity_result_value(value: &Value) -> Option<ClosedIssueEvidence> {
-    value
-        .get("signals")
-        .and_then(Value::as_array)
-        .into_iter()
-        .flatten()
-        .find_map(|signal| {
-            let signal_type = signal.get("signal_type").and_then(non_empty_json_string)?;
-            match signal_type.as_str() {
-                ISSUE_CLOSED_SIGNAL | ISSUE_ALREADY_RESOLVED_SIGNAL => signal
-                    .get("signal")
-                    .and_then(|payload| closed_issue_evidence_from_value(payload, &signal_type)),
-                _ => None,
-            }
-        })
-        .or_else(|| {
-            value
-                .get("artifacts")
-                .and_then(Value::as_array)
-                .into_iter()
-                .flatten()
-                .filter(|artifact| {
-                    artifact.get("artifact_type").and_then(Value::as_str)
-                        == Some(ISSUE_STATE_ARTIFACT)
-                })
-                .find_map(|artifact| {
-                    artifact
-                        .get("artifact")
-                        .filter(|payload| issue_state_is_closed(payload))
-                        .and_then(|payload| {
-                            closed_issue_evidence_from_value(payload, ISSUE_STATE_ARTIFACT)
-                        })
-                })
-        })
-}
-
-fn closed_issue_evidence_from_value(value: &Value, source: &str) -> Option<ClosedIssueEvidence> {
-    let issue_number = value.get("issue_number").and_then(Value::as_u64);
-    let issue_url = value
-        .get("issue_url")
-        .or_else(|| value.get("html_url"))
-        .or_else(|| value.get("url"))
-        .and_then(non_empty_json_string);
-    let state = value.get("state").and_then(non_empty_json_string);
-    let closed = issue_state_is_closed(value);
-
-    if !closed || (issue_number.is_none() && issue_url.is_none()) {
-        return None;
-    }
-
-    let mut facts = vec![format!("source={source}")];
-    if let Some(issue_number) = issue_number {
-        facts.push(format!("issue_number={issue_number}"));
-    }
-    if let Some(issue_url) = issue_url.clone() {
-        facts.push(format!("issue_url={issue_url}"));
-    }
-    if let Some(state) = state.clone() {
-        facts.push(format!("state={state}"));
-    }
-    if closed {
-        facts.push("closed=true".to_string());
-    }
-
-    Some(ClosedIssueEvidence {
-        summary: facts.join(" "),
-        payload: json!({
-            "source": source,
-            "issue_number": issue_number,
-            "issue_url": issue_url,
-            "state": state,
-            "closed": closed,
-        }),
-    })
-}
-
-fn issue_state_is_closed(value: &Value) -> bool {
-    value
-        .get("closed")
-        .and_then(Value::as_bool)
-        .unwrap_or(false)
-        || value
-            .get("state")
-            .and_then(Value::as_str)
-            .is_some_and(|state| {
-                state.trim().eq_ignore_ascii_case("closed")
-                    || state.trim().eq_ignore_ascii_case("resolved")
-            })
-        || value
-            .get("is_closed")
-            .and_then(Value::as_bool)
-            .unwrap_or(false)
-}
-
-fn pr_feedback_sweep_decision_from_activity_result(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> Option<WorkflowDecision> {
-    if instance.definition_id != GITHUB_ISSUE_PR_DEFINITION_ID
-        || !matches!(
-            result.activity.as_str(),
-            "sweep_pr_feedback" | PR_FEEDBACK_INSPECT_ACTIVITY
-        )
-    {
-        return None;
-    }
-    if !matches!(
-        instance.state.as_str(),
-        "pr_open" | "awaiting_feedback" | "addressing_feedback"
-    ) {
-        return None;
-    }
-    let outcome = pr_feedback_outcome_from_signals(result)?;
-    let pr_number = result_signal_u64(result, "pr_number").or_else(|| {
-        instance
-            .data
-            .get("pr_number")
-            .and_then(|value| value.as_u64())
-    })?;
-    let pr_url =
-        result_signal_string(result, "pr_url").or_else(|| optional_data_string(instance, "pr_url"));
-    let task_id = event_field_string(event, "runtime_job_id")
-        .or_else(|| optional_data_string(instance, "task_id"))
-        .unwrap_or_else(|| event.id.clone());
-    Some(
-        build_pr_feedback_decision(
-            instance,
-            PrFeedbackDecisionInput {
-                task_id: &task_id,
-                pr_number,
-                pr_url: pr_url.as_deref(),
-                outcome,
-                summary: result.summary.as_str(),
-            },
-        )
-        .decision
-        .with_evidence(runtime_completion_evidence(event, result)),
-    )
-}
-
-fn pr_feedback_child_decision_from_activity_result(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> Option<WorkflowDecision> {
-    if (
-        instance.definition_id.as_str(),
-        instance.state.as_str(),
-        result.activity.as_str(),
-    ) != (
-        PR_FEEDBACK_DEFINITION_ID,
-        "inspecting",
-        PR_FEEDBACK_INSPECT_ACTIVITY,
-    ) {
-        return None;
-    }
-
-    let outcome = pr_feedback_outcome_from_signals(result)?;
-    let (next_state, decision, reason) = match outcome {
-        PrFeedbackOutcome::BlockingFeedback => (
-            "feedback_found",
-            "record_feedback_found",
-            "PR feedback inspection found actionable feedback.",
-        ),
-        PrFeedbackOutcome::NoActionableFeedback => (
-            "no_actionable_feedback",
-            "record_no_actionable_feedback",
-            "PR feedback inspection found no actionable feedback.",
-        ),
-        PrFeedbackOutcome::ReadyToMerge => (
-            "ready_to_merge",
-            "record_ready_to_merge",
-            "PR feedback inspection found the PR ready to merge.",
-        ),
-    };
-
-    Some(
-        WorkflowDecision::new(&instance.id, &instance.state, decision, next_state, reason)
-            .with_evidence(runtime_completion_evidence(event, result))
-            .high_confidence(),
-    )
-}
-
-fn pr_feedback_outcome_from_signals(result: &ActivityResult) -> Option<PrFeedbackOutcome> {
-    if has_signal(result, "FeedbackFound")
-        || has_signal(result, "ChangesRequested")
-        || has_signal(result, "ChecksFailed")
-    {
-        return Some(PrFeedbackOutcome::BlockingFeedback);
-    }
-    if has_signal(result, "PrReadyToMerge") {
-        return Some(PrFeedbackOutcome::ReadyToMerge);
-    }
-    if has_signal(result, "NoFeedbackFound") {
-        return Some(PrFeedbackOutcome::NoActionableFeedback);
-    }
-    None
-}
-
-fn has_signal(result: &ActivityResult, signal_type: &str) -> bool {
-    result
-        .signals
-        .iter()
-        .any(|signal| signal.signal_type == signal_type)
-}
-
-fn result_signal_u64(result: &ActivityResult, field: &str) -> Option<u64> {
-    result
-        .signals
-        .iter()
-        .find_map(|signal| signal.signal.get(field).and_then(json_value_u64))
-}
-
-fn result_signal_string(result: &ActivityResult, field: &str) -> Option<String> {
-    result.signals.iter().find_map(|signal| {
-        signal
-            .signal
-            .get(field)
-            .and_then(|value| value.as_str())
-            .filter(|value| !value.trim().is_empty())
-            .map(ToOwned::to_owned)
-    })
-}
-
-fn optional_data_string(instance: &WorkflowInstance, field: &str) -> Option<String> {
-    instance
-        .data
-        .get(field)
-        .and_then(|value| value.as_str())
-        .filter(|value| !value.trim().is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn json_value_u64(value: &Value) -> Option<u64> {
-    value
-        .as_u64()
-        .or_else(|| value.as_str().and_then(|raw| raw.parse::<u64>().ok()))
-}
-
-fn non_empty_json_string(value: &Value) -> Option<String> {
-    value
-        .as_str()
-        .filter(|value| !value.trim().is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn string_array_field(value: &Value, field: &str) -> Vec<String> {
-    value
-        .get(field)
-        .and_then(Value::as_array)
-        .map(|values| {
-            values
-                .iter()
-                .filter_map(non_empty_json_string)
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default()
-}
-
-fn u64_array_field(value: &Value, field: &str) -> Vec<u64> {
-    value
-        .get(field)
-        .and_then(Value::as_array)
-        .map(|values| values.iter().filter_map(json_value_u64).collect::<Vec<_>>())
-        .unwrap_or_default()
-}
-
-fn array_items<'a>(value: &'a Value, field: &str) -> Vec<&'a Value> {
-    value
-        .get(field)
-        .and_then(Value::as_array)
-        .map(|items| items.iter().collect())
-        .unwrap_or_default()
-}
-
-fn runtime_blocked_decision(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> WorkflowDecision {
-    let reason = runtime_failure_reason(result, "Runtime activity was blocked.");
-    WorkflowDecision::new(
-        &instance.id,
-        &instance.state,
-        "block_after_runtime_activity",
-        "blocked",
-        &reason,
-    )
-    .with_command(WorkflowCommand::mark_blocked(
-        &reason,
-        format!("runtime-completion:{}:blocked", event.id),
-    ))
-    .with_evidence(runtime_completion_evidence(event, result))
-    .high_confidence()
-}
-
-fn runtime_failed_decision(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> WorkflowDecision {
-    let reason = runtime_failure_reason(result, "Runtime activity failed.");
-    let mut command_payload = json!({ "reason": reason.as_str() });
-    if let Some(error_kind) = result.error_kind {
-        if let Some(object) = command_payload.as_object_mut() {
-            object.insert("error_kind".to_string(), json!(error_kind));
-        }
-    }
-    WorkflowDecision::new(
-        &instance.id,
-        &instance.state,
-        "fail_after_runtime_activity",
-        "failed",
-        &reason,
-    )
-    .with_command(WorkflowCommand::new(
-        WorkflowCommandType::MarkFailed,
-        format!("runtime-completion:{}:failed", event.id),
-        command_payload,
-    ))
-    .with_evidence(runtime_completion_evidence(event, result))
-    .high_confidence()
-}
-
-fn retry_failed_activity_decision(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> Option<WorkflowDecision> {
-    if !is_retryable_error_kind(result.error_kind) {
-        return None;
-    }
-    if !supports_same_state_activity_retry(&instance.definition_id, &instance.state) {
-        return None;
-    }
-    let retry_attempt = failed_activity_retry_attempt(event);
-    let activity = retry_activity_name(event, result)?;
-    let retry_limit = failed_activity_retry_limit(instance, &activity)?;
-    if retry_attempt >= retry_limit {
-        return None;
-    }
-    let next_attempt = retry_attempt + 1;
-    let reason = runtime_failure_reason(result, "Runtime activity failed.");
-    let retry_schedule = failed_activity_retry_schedule(instance, &activity, next_attempt);
-    Some(
-        WorkflowDecision::new(
-            &instance.id,
-            &instance.state,
-            "retry_failed_runtime_activity",
-            &instance.state,
-            format!(
-                "Runtime activity `{activity}` failed; retrying attempt {next_attempt} of {retry_limit}. Last error: {reason}"
-            ),
-        )
-        .with_command(retry_command(
-            event,
-            &activity,
-            result.error_kind,
-            next_attempt,
-            retry_limit,
-            &reason,
-            retry_schedule,
-        ))
-        .with_evidence(runtime_completion_evidence(event, result))
-        .high_confidence(),
-    )
-}
-
-fn is_retryable_error_kind(error_kind: Option<ActivityErrorKind>) -> bool {
-    !matches!(
-        error_kind,
-        Some(ActivityErrorKind::Fatal | ActivityErrorKind::Configuration)
-    )
-}
-
-fn runtime_cancelled_decision(
-    instance: &WorkflowInstance,
-    event: &WorkflowEvent,
-    result: &ActivityResult,
-) -> WorkflowDecision {
-    let reason = runtime_failure_reason(result, "Runtime activity was cancelled.");
-    WorkflowDecision::new(
-        &instance.id,
-        &instance.state,
-        "cancel_after_runtime_activity",
-        "cancelled",
-        &reason,
-    )
-    .with_command(WorkflowCommand::new(
-        WorkflowCommandType::MarkCancelled,
-        format!("runtime-completion:{}:cancelled", event.id),
-        json!({ "reason": reason }),
-    ))
-    .with_evidence(runtime_completion_evidence(event, result))
-    .high_confidence()
-}
-
-fn runtime_failure_reason(result: &ActivityResult, fallback: &str) -> String {
-    result
-        .error
-        .as_deref()
-        .filter(|value| !value.trim().is_empty())
-        .or_else(|| (!result.summary.trim().is_empty()).then_some(result.summary.trim()))
-        .unwrap_or(fallback)
-        .to_string()
-}
-
-fn failed_activity_retry_limit(instance: &WorkflowInstance, activity: &str) -> Option<u64> {
-    if let Some(limit) = retry_policy_u64(instance, activity, "max_failed_activity_retries") {
-        return (limit > 0).then_some(limit);
-    }
-    None
-}
-
-#[derive(Debug, Clone, Copy)]
-struct RetrySchedule {
-    delay_secs: u64,
-    not_before: DateTime<Utc>,
-}
-
-fn failed_activity_retry_schedule(
-    instance: &WorkflowInstance,
-    activity: &str,
-    next_attempt: u64,
-) -> Option<RetrySchedule> {
-    let base_delay = retry_policy_u64(instance, activity, "retry_delay_secs")?;
-    if base_delay == 0 {
-        return None;
-    }
-    let multiplier = 1_u64
-        .checked_shl(next_attempt.saturating_sub(1).min(63) as u32)
-        .unwrap_or(u64::MAX);
-    let mut delay_secs = base_delay.saturating_mul(multiplier);
-    if let Some(max_delay) = retry_policy_u64(instance, activity, "max_retry_delay_secs")
-        .filter(|max_delay| *max_delay > 0)
-    {
-        delay_secs = delay_secs.min(max_delay);
-    }
-    const MAX_SAFE_RETRY_DELAY_SECS: u64 = 30 * 24 * 60 * 60;
-    delay_secs = delay_secs.min(MAX_SAFE_RETRY_DELAY_SECS);
-    Some(RetrySchedule {
-        delay_secs,
-        not_before: Utc::now() + Duration::seconds(delay_secs as i64),
-    })
-}
-
-fn retry_policy_u64(instance: &WorkflowInstance, activity: &str, field: &str) -> Option<u64> {
-    let policy = instance.data.get("runtime_retry_policy")?;
-    policy
-        .get("activity_retries")
-        .and_then(|activities| activities.get(activity))
-        .and_then(|activity_policy| activity_policy.get(field))
-        .and_then(Value::as_u64)
-        .or_else(|| policy.get(field).and_then(Value::as_u64))
-}
-
-fn failed_activity_retry_attempt(event: &WorkflowEvent) -> u64 {
-    event
-        .event
-        .get("command")
-        .and_then(|command| command.get("command"))
-        .and_then(|command| command.get("retry_attempt"))
-        .and_then(|value| value.as_u64())
-        .unwrap_or(0)
-}
-
-fn retry_activity_name(event: &WorkflowEvent, result: &ActivityResult) -> Option<String> {
-    event
-        .event
-        .get("command")
-        .and_then(|command| command.get("command"))
-        .and_then(|command| command.get("activity"))
-        .and_then(|value| value.as_str())
-        .filter(|activity| !activity.trim().is_empty())
-        .or_else(|| event_command_type(event))
-        .or_else(|| (!result.activity.trim().is_empty()).then_some(result.activity.as_str()))
-        .map(str::to_string)
-}
-
-fn retry_command(
-    event: &WorkflowEvent,
-    activity: &str,
-    error_kind: Option<ActivityErrorKind>,
-    next_attempt: u64,
-    retry_limit: u64,
-    reason: &str,
-    retry_schedule: Option<RetrySchedule>,
-) -> WorkflowCommand {
-    let previous = event_workflow_command(event);
-    let command_type = previous
-        .as_ref()
-        .map(|command| command.command_type)
-        .unwrap_or(WorkflowCommandType::EnqueueActivity);
-    let mut command_payload = previous
-        .map(|command| command.command)
-        .unwrap_or_else(|| json!({ "activity": activity }));
-    if let Some(object) = command_payload.as_object_mut() {
-        if command_type == WorkflowCommandType::EnqueueActivity && !object.contains_key("activity")
-        {
-            object.insert("activity".to_string(), json!(activity));
-        }
-        object.insert("retry_attempt".to_string(), json!(next_attempt));
-        object.insert(
-            "max_failed_activity_retries".to_string(),
-            json!(retry_limit),
-        );
-        object.insert(
-            "previous_command_id".to_string(),
-            optional_json_string(event_field_string(event, "command_id")),
-        );
-        object.insert(
-            "previous_runtime_job_id".to_string(),
-            optional_json_string(event_field_string(event, "runtime_job_id")),
-        );
-        object.insert("previous_error".to_string(), json!(reason));
-        if let Some(error_kind) = error_kind {
-            object.insert("previous_error_kind".to_string(), json!(error_kind));
-        }
-        if let Some(schedule) = retry_schedule {
-            object.insert("retry_delay_secs".to_string(), json!(schedule.delay_secs));
-            object.insert(
-                "retry_not_before".to_string(),
-                json!(schedule.not_before.to_rfc3339()),
-            );
-        }
-    } else {
-        command_payload = json!({
-            "activity": activity,
-            "retry_attempt": next_attempt,
-            "max_failed_activity_retries": retry_limit,
-            "previous_command_id": event_field_string(event, "command_id"),
-            "previous_runtime_job_id": event_field_string(event, "runtime_job_id"),
-            "previous_error": reason,
-        });
-        if let Some(error_kind) = error_kind {
-            if let Some(object) = command_payload.as_object_mut() {
-                object.insert("previous_error_kind".to_string(), json!(error_kind));
-            }
-        }
-        if let Some(schedule) = retry_schedule {
-            if let Some(object) = command_payload.as_object_mut() {
-                object.insert("retry_delay_secs".to_string(), json!(schedule.delay_secs));
-                object.insert(
-                    "retry_not_before".to_string(),
-                    json!(schedule.not_before.to_rfc3339()),
-                );
-            }
-        }
-    }
-    WorkflowCommand::new(
-        command_type,
-        format!(
-            "runtime-completion:{}:retry:{}:{}",
-            event.id, activity, next_attempt
-        ),
-        command_payload,
-    )
-}
-
-fn supports_same_state_activity_retry(definition_id: &str, state: &str) -> bool {
-    matches!(
-        (definition_id, state),
-        (
-            GITHUB_ISSUE_PR_DEFINITION_ID,
-            "implementing" | "awaiting_feedback" | "addressing_feedback"
-        ) | (
-            REPO_BACKLOG_DEFINITION_ID,
-            "scanning" | "dispatching" | "reconciling" | "planning_batch"
-        ) | (PR_FEEDBACK_DEFINITION_ID, "inspecting")
-            | (PROMPT_TASK_DEFINITION_ID, "implementing")
-            | (QUALITY_GATE_DEFINITION_ID, "checking")
-    )
-}
-
-fn event_field_string(event: &WorkflowEvent, field: &str) -> Option<String> {
-    event
-        .event
-        .get(field)
-        .and_then(|value| value.as_str())
-        .map(str::to_string)
-}
-
-fn event_command_type(event: &WorkflowEvent) -> Option<&str> {
-    event
-        .event
-        .get("command")
-        .and_then(|command| command.get("command_type"))
-        .and_then(|value| value.as_str())
-}
-
-fn event_workflow_command(event: &WorkflowEvent) -> Option<WorkflowCommand> {
-    event
-        .event
-        .get("command")
-        .cloned()
-        .and_then(|value| serde_json::from_value(value).ok())
-}
-
-fn optional_json_string(value: Option<String>) -> Value {
-    value.map(Value::String).unwrap_or(Value::Null)
-}
-
-fn runtime_completion_evidence(event: &WorkflowEvent, result: &ActivityResult) -> WorkflowEvidence {
-    let command_id = event
-        .event
-        .get("command_id")
-        .and_then(|value| value.as_str())
-        .unwrap_or("<unknown>");
-    let runtime_job_id = event
-        .event
-        .get("runtime_job_id")
-        .and_then(|value| value.as_str())
-        .unwrap_or("<unknown>");
-    WorkflowEvidence::new(
-        "runtime_completion",
-        format!(
-            "activity={} status={:?} command_id={} runtime_job_id={} summary={}",
-            result.activity, result.status, command_id, runtime_job_id, result.summary
-        ),
-    )
 }
